requirements.txt
streamlit
Pillow
numpy
easyocr